Intorduction

Intorduction

As we know triangular meshes can be stored in a computer in a multiple different ways, each having strength and weaknesses in a particular cases. Not always it is clear which datastructure would be most performant for a numerical algorithm so it is wise to write a datastructure generic code. This package is exactly for that purpose for orriented closed surfaces.

The simplest representation of triangular mesh topology is in array Array{Faces{3,Int},1} containing a list of triangular faces which are defined by their vertices. Often in a numerical code one wants not only to iterate over faces or vertices, but also in case of singularity substraction, integration and local property estimation like normal vectors and curvature to know what are neighbouring verticies surrounding a given vertex while keeping track of orrientation of the normals. Also one wishes to modify the topology itself by colloapsing, flipingn and splitting edges. This is why different datstructrures are needed for different problems.

Fortunatelly it is possible to abstract queries of the topology through iterators:

SurfaceTopology.Faces — Function.

Faces(t) returns an iterator for faces from representation of topology t

source

SurfaceTopology.Edges — Function.

Edges(t) returns an iterator for edges from representation of topology t

source

and circulators:

SurfaceTopology.VertexRing — Type.

VertexRing(v,t)

Construct a vertex ring iterator at vertex v from a given topology t.

source

SurfaceTopology.EdgeRing — Type.

EdgeRing(v,t)

Construct an edge ring iterator at vertex v from a given topology t.

source

API

At the moment the package implements three different kinds of datastructures. The simplest one is PlainDS one which stores list of faces and is just an alias to Array{Faces{3,Int},1}. As an example how taht works let's deffine the datastructure:

using GeometryTypes
using SurfaceTopology

faces = Face{3,Int64}[
    [1, 12, 6], [1, 6, 2], [1, 2, 8], [1, 8, 11], [1, 11, 12], [2, 6, 10], [6, 12, 5],
    [12, 11, 3], [11, 8, 7], [8, 2, 9], [4, 10, 5], [4, 5, 3], [4, 3, 7], [4, 7, 9],
    [4, 9, 10], [5, 10, 6], [3, 5, 12], [7, 3, 11], [9, 7, 8], [10, 9, 2]
]

20-element Array{GeometryTypes.Face{3,Int64},1}:
 [1, 12, 6] 
 [1, 6, 2]  
 [1, 2, 8]  
 [1, 8, 11] 
 [1, 11, 12]
 [2, 6, 10] 
 [6, 12, 5] 
 [12, 11, 3]
 [11, 8, 7] 
 [8, 2, 9]  
 [4, 10, 5] 
 [4, 5, 3]  
 [4, 3, 7]  
 [4, 7, 9]  
 [4, 9, 10] 
 [5, 10, 6] 
 [3, 5, 12] 
 [7, 3, 11] 
 [9, 7, 8]  
 [10, 9, 2] 

The datastructure faces can be directly used for the queries. The iterators can be executed:

collect(Faces(faces))

20-element Array{GeometryTypes.Face{3,Int64},1}:
 [1, 12, 6] 
 [1, 6, 2]  
 [1, 2, 8]  
 [1, 8, 11] 
 [1, 11, 12]
 [2, 6, 10] 
 [6, 12, 5] 
 [12, 11, 3]
 [11, 8, 7] 
 [8, 2, 9]  
 [4, 10, 5] 
 [4, 5, 3]  
 [4, 3, 7]  
 [4, 7, 9]  
 [4, 9, 10] 
 [5, 10, 6] 
 [3, 5, 12] 
 [7, 3, 11] 
 [9, 7, 8]  
 [10, 9, 2] 

and

collect(Edges(faces))

30-element Array{GeometryTypes.Face{2,Int64},1}:
 [1, 12] 
 [1, 6]  
 [1, 2]  
 [2, 8]  
 [1, 8]  
 [8, 11] 
 [1, 11] 
 [11, 12]
 [2, 6]  
 [6, 10] 
 ⋮       
 [4, 9]  
 [9, 10] 
 [5, 10] 
 [3, 5]  
 [5, 12] 
 [3, 11] 
 [7, 8]  
 [8, 9]  
 [2, 10] 

giving us desirable output.

The circulators which for this simple datastructure requires to do a full lookup on the array can simply be executed as:

collect(VertexRing(3,faces))

5-element Array{Int64,1}:
 12
  4
  7
  5
 11

and

collect(EdgeRing(3,faces))

5-element Array{GeometryTypes.Face{2,Int64},1}:
 [12, 11]
 [4, 5]  
 [7, 4]  
 [5, 12] 
 [11, 7] 

In practice one should use EdgeRing over VertexRing since in the later one vertices are not ordered.

Datastructures

The same API works for all other datastructures. There is a datastructure CachedDS built on top of PlainDS stores closest vertices (vertex ring). Then there is a datastructure FaceDS which with PlainDS also stores neighbouring faces which have a common edge. And then there are most popular datastructures HalfEdgeDS (implemented as EdgeDS).

The simplest extension is just a plain caching implemented under CacheDS which for each vertex stores it's surrounding verticies.

SurfaceTopology.CachedDS — Type.

A datastructure which in addition to a list of faces stores connectivity information for each vertex.

source

SurfaceTopology.CachedDS — Method.

CacheDS(t)

Constructs cached face based datastructure CachedDS from arbitrary topology t which provides EdgeRing iterator.

source

which can be initialised from PlainDS

cachedtopology = CachedDS(faces)

CachedDS(GeometryTypes.Face{3,Int64}[[1, 12, 6], [1, 6, 2], [1, 2, 8], [1, 8, 11], [1, 11, 12], [2, 6, 10], [6, 12, 5], [12, 11, 3], [11, 8, 7], [8, 2, 9], [4, 10, 5], [4, 5, 3], [4, 3, 7], [4, 7, 9], [4, 9, 10], [5, 10, 6], [3, 5, 12], [7, 3, 11], [9, 7, 8], [10, 9, 2]], Array{Int64,1}[[6, 2, 8, 11, 12], [6, 10, 9, 8, 1], [11, 7, 4, 5, 12], [5, 3, 7, 9, 10], [12, 3, 4, 10, 6], [12, 5, 10, 2, 1], [8, 9, 4, 3, 11], [2, 9, 7, 11, 1], [2, 10, 4, 7, 8], [6, 5, 4, 9, 2], [8, 7, 3, 12, 1], [1, 11, 3, 5, 6]])

And the same API can be used for querries:

collect(VertexRing(3,cachedtopology))

5-element Array{Int64,1}:
 11
  7
  4
  5
 12

A more advanced datastructure is a face based datastructure which in this library is defined as FaceDS which additionally for each face stores 3 neigbouring face indicies.

SurfaceTopology.FaceDS — Type.

A face based datastructure storing faces, neighbour face indices and vertex to face map arrays.

source

SurfaceTopology.FaceDS — Method.

FaceDS(faces::PlainDS)

Constructs a face based datastructure from PlainDS. Returns a struct FaceDS with original faces, computed neighbour faces and vertex to face map (one face for each vertex).

source

which again can be initialised from PlainDS

facedstopology = FaceDS(faces)

FaceDS(GeometryTypes.Face{3,Int64}[[1, 12, 6], [1, 6, 2], [1, 2, 8], [1, 8, 11], [1, 11, 12], [2, 6, 10], [6, 12, 5], [12, 11, 3], [11, 8, 7], [8, 2, 9], [4, 10, 5], [4, 5, 3], [4, 3, 7], [4, 7, 9], [4, 9, 10], [5, 10, 6], [3, 5, 12], [7, 3, 11], [9, 7, 8], [10, 9, 2]], GeometryTypes.Face{3,Int64}[[7, 2, 5], [6, 3, 1], [10, 4, 2], [9, 5, 3], [8, 1, 4], [16, 20, 2], [17, 16, 1], [18, 17, 5], [19, 18, 4], [20, 19, 3], [16, 12, 15], [17, 13, 11], [18, 14, 12], [19, 15, 13], [20, 11, 14], [6, 7, 11], [7, 8, 12], [8, 9, 13], [9, 10, 14], [10, 6, 15]], [1, 2, 8, 11, 7, 1, 9, 3, 10, 6, 4, 1])

and what would one expect

collect(VertexRing(3,facedstopology))

5-element Array{Int64,1}:
 12
 11
  7
  4
  5

works.

The last in the list is half edge datastructure EdgeDS which stores edges with three numbers - base vertex index, next edge index and twin edge index.

SurfaceTopology.EdgeDS — Type.

EdgeDS(faces::PlainDS)

Constructs and returns edge based datastructure EdgeDS from plain face based datastructure PlainDS. Half edge based datastructure EdgeDS stores list of edges consisting of base vertex, next edge index and twin edge index.

source

To initiate this datastructure one executes:

edgedstopology = EdgeDS(faces)

EdgeDS(GeometryTypes.Face{3,Int64}[[1, 2, 15], [12, 3, 19], [6, 1, 4], [1, 5, 3], [6, 6, 16], [2, 4, 7], [1, 8, 6], [2, 9, 28], [8, 7, 10], [1, 11, 9]  …  [12, 49, 24], [7, 53, 38], [3, 54, 23], [11, 52, 27], [9, 56, 41], [7, 57, 26], [8, 55, 30], [10, 59, 44], [9, 60, 29], [2, 58, 18]])

collect(VertexRing(3,edgedstopology))

5-element Array{Int64,1}:
 12
 11
  7
  4
  5

Wishlist

At the moment the package is able to only answer queries, but it would be desirable to also be able to do topological operations. For completition thoose would include:

edgeflip(topology,edge)

edgesplit(topology,edge)

edgecollapse(topology,edge)

And with them also a method for defragmenting the topology. Unfortunatelly due to irrelevance of this package for my present research, the development of that on my own will be slow. I hope that clarity and simplicity of this package could serve someone as a first step and so eventually topological operations would be impemented out of necessity.